Split out identifyUser from lib/analytics
